backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.10
pydantic>=2.10.4
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
email-validator>=2.1.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart
PyYAML>=6.0

